Add explicit dependency on CocoaPods w/in main scheme
This fixes the xctool build failure when linking CocoaPods.

The problem was xctool was not detecting the implicit CocoaPods
dependency. The workaround[1] is to disable implicit dependencies and
explicitly add the main Pods target (and Pods-test target, if one
exists) as build dependencies within your scheme *before* any remaining
targets.

[1]: https://github.com/facebook/xctool/issues/16#issuecomment-17444311